feat : Add logger and quick fix. - Add winstom package for logging - Add util folder to encapsulate the logger file. - Refactor the DiscussionItem to include the userId
