+++ project-6/capstone/project-dami-backend/package.json
@@ -9,7 +9,8 @@
   },
   "dependencies": {
     "source-map-support": "^0.5.10",
-    "uuid": "^8.3.1"
+    "uuid": "^8.3.1",
+    "winston": "^3.3.3"
   },
   "devDependencies": {
     "@types/aws-lambda": "^8.10.17",
