+++ project-6/capstone/project-dami-backend/serverless.yml
@@ -1,4 +1,4 @@
-service: dami-sls-api
+service: dami-discussion-service
 
 # Add the serverless-webpack plugin
 plugins:
@@ -19,7 +19,7 @@ provider:
     minimumCompressionSize: 1024 # Enable gzip compression for responses > 1 KB
   environment:
     AWS_NODEJS_CONNECTION_REUSE_ENABLED: 1
-    DISCUSSION_TABLE : Discussion-${opt:stage, 'dev'}
+    DISCUSSION_TABLE: Discussion-${opt:stage, 'dev'}
 
 custom:
   webpack:
@@ -42,24 +42,23 @@ functions:
     name: ${self:service}-createDiscussion-${opt:stage, self:provider.stage, 'dev'}
     handler: src/lambda/http/createDiscussion.handler
     description: Create discussion lambda
-    package: {}
     events:
       - http:
           method: post
           path: discussion
           cors: true
           reqValidatorName: RequestBodyValidator
-          documentation:  
+          documentation:
             summary: Create a new discussion
             description: Create a new discussion
             requestModels:
-              "application/json": DiscussionRequest
+              'application/json': DiscussionRequest
     iamRoleStatementsName: createDiscussion-${self:provider.stage}-lambdaRole
     iamRoleStatements:
       - Effect: Allow
         Action:
-        - dynamodb:Scan
-        - dynamodb:PutItem
+          - dynamodb:Scan
+          - dynamodb:PutItem
         Resource: arn:aws:dynamodb:${self:provider.region}:*:table/${self:provider.environment.DISCUSSION_TABLE}
 
 resources:
@@ -67,26 +66,24 @@ resources:
     RequestBodyValidator:
       Type: AWS::ApiGateway::RequestValidator
       Properties:
-        Name: "request-body-validator"
+        Name: 'request-body-validator'
         RestApiId:
           Ref: ApiGatewayRestApi
         ValidateRequestBody: true
         ValidateRequestParameters: false
 
     DiscussionTable:
-    Type: AWS::DynamoDB::Table
-    Properties:
-      AttributeDefinitions:
-        - AttributeName: userId
-          AttributeType: S
-        - AttributeName: createdAt
-          AttributeType: S
-        - AttributeName: discussionId
-          AttributeType: S
-      KeySchema:
-        - AttributeName: userId
-          KeyType: HASH
-        - AttributeName: discussionId
-          KeyType: RANGE
-      BillingMode: PAY_PER_REQUEST
-      TableName: ${self:provider.environment.DISCUSSION_TABLE}
+      Type: AWS::DynamoDB::Table
+      Properties:
+        AttributeDefinitions:
+          - AttributeName: userId
+            AttributeType: S
+          - AttributeName: discussionId
+            AttributeType: S
+        KeySchema:
+          - AttributeName: userId
+            KeyType: HASH
+          - AttributeName: discussionId
+            KeyType: RANGE
+        BillingMode: PAY_PER_REQUEST
+        TableName: ${self:provider.environment.DISCUSSION_TABLE}
