+++ project-6/capstone/project-dami-backend/src/lambda/http/createDiscussion.ts
@@ -1,18 +1,25 @@
-import { APIGatewayProxyHandler } from 'aws-lambda'
+import { APIGatewayProxyEvent, APIGatewayProxyHandler } from 'aws-lambda'
 import 'source-map-support/register'
 import { onCreateDiscussion } from '../../businessLogic/discussion'
 import { DiscussionRequest } from '../../requests/DiscussionRequest'
+import { createLogger } from '../../utils/logger'
 
-export const handler: APIGatewayProxyHandler = async (event, _context) => {
+export const handler: APIGatewayProxyHandler = async (
+  event: APIGatewayProxyEvent,
+  _context
+) => {
+  const logger = createLogger('create-discussion')
   const newDiscussion: DiscussionRequest = JSON.parse(event.body)
-  const newItem = await onCreateDiscussion(newDiscussion)
+  const item = await onCreateDiscussion(newDiscussion)
+
+  logger.info('DISCUSSION : ', item)
   return {
     statusCode: 201,
     headers: {
       'Access-Control-Allow-Origin': '*',
     },
     body: JSON.stringify({
-      newItem,
+      item,
     }),
   }
 }
