+++ project-6/capstone/project-dami-backend/src/models/DiscussionItem.ts
@@ -3,4 +3,5 @@ export interface DiscussionItem {
   shortDescription: string
   createdAt: string
   discussionId: string
+  userId: string
 }
