+++ project-6/capstone/project-dami-backend/src/utils/logger.ts
@@ -1,0 +1,15 @@
+import * as winston from 'winston'
+
+/**
+ * Create a logger instance to write log messages in JSON format.
+ *
+ * @param loggerName - a name of a logger that will be added to all messages
+ */
+export function createLogger(loggerName: string) {
+  return winston.createLogger({
+    level: 'info',
+    format: winston.format.json(),
+    defaultMeta: { name: loggerName },
+    transports: [new winston.transports.Console()],
+  })
+}
